Moves Checked code under detail
