+++ Library/include/playrho/Finite.hpp
@@ -21,41 +21,19 @@
 #ifndef PLAYRHO_FINITE_HPP
 #define PLAYRHO_FINITE_HPP
 
-#include <playrho/Checked.hpp>
-#include <playrho/Math.hpp> // for playrho::isfinite
+#include <playrho/detail/FiniteChecker.hpp>
 
 namespace playrho {
-
-/// @brief Finite constrained value checker.
-template <typename T>
-struct FiniteChecker {
-
-    /// @brief Default value supplying functor.
-    constexpr auto operator()() noexcept(noexcept(static_cast<T>(0))) -> decltype(static_cast<T>(0))
-    {
-        return static_cast<T>(0);
-    }
-
-    /// @brief Value checking functor.
-    auto operator()(const T& v) noexcept(noexcept(isfinite(v)))
-        -> decltype(isfinite(v), static_cast<const char*>(nullptr))
-    {
-        if (!isfinite(v)) {
-            return "value not finite";
-        }
-        return {};
-    }
-};
 
 /// @ingroup CheckedTypes
 /// @brief Finite constrained value type.
 template <typename T>
-using Finite = Checked<T, FiniteChecker<T>>;
+using Finite = detail::Checked<T, detail::FiniteChecker<T>>;
 
 /// @ingroup CheckedTypes
 /// @brief Fast failing finite constrained value type.
 template <typename T>
-using FiniteFF = Checked<T, FiniteChecker<T>, true>;
+using FiniteFF = detail::Checked<T, detail::FiniteChecker<T>, true>;
 
 static_assert(std::is_default_constructible<Finite<int>>::value);
