+++ Library/include/playrho/NonZero.hpp
@@ -21,46 +21,31 @@
 #ifndef PLAYRHO_NONZERO_HPP
 #define PLAYRHO_NONZERO_HPP
 
-#include <playrho/Checked.hpp>
+#include <playrho/detail/NonZeroChecker.hpp>
 
 namespace playrho {
-
-/// @brief Non-zero constrained value checker.
-template <typename T>
-struct NonZeroChecker {
-
-    /// @brief Value checking functor.
-    constexpr auto operator()(const T& v) noexcept
-        -> decltype(!(v != static_cast<T>(0)), static_cast<const char*>(nullptr))
-    {
-        if (!(v != static_cast<T>(0))) {
-            return "value not non-zero";
-        }
-        return {};
-    }
-};
 
 /// @ingroup CheckedTypes
 /// @brief Non-zero constrained value type.
 template <typename T>
-using NonZero = std::enable_if_t<!std::is_pointer<T>::value, Checked<T, NonZeroChecker<T>>>;
+using NonZero = std::enable_if_t<!std::is_pointer<T>::value, detail::Checked<T, detail::NonZeroChecker<T>>>;
 
 /// @ingroup CheckedTypes
 /// @brief Fast failing non-zero constrained value type.
 template <typename T>
-using NonZeroFF = std::enable_if_t<!std::is_pointer<T>::value, Checked<T, NonZeroChecker<T>, true>>;
+using NonZeroFF = std::enable_if_t<!std::is_pointer<T>::value, detail::Checked<T, detail::NonZeroChecker<T>, true>>;
 
 static_assert(!std::is_default_constructible<NonZero<int>>::value);
 
 /// @ingroup CheckedTypes
 /// @brief Non-null constrained value type.
 template <typename T>
-using NonNull = std::enable_if_t<std::is_pointer<T>::value, Checked<T, NonZeroChecker<T>>>;
+using NonNull = std::enable_if_t<std::is_pointer<T>::value, detail::Checked<T, detail::NonZeroChecker<T>>>;
 
 /// @ingroup CheckedTypes
 /// @brief Fast failing non-null constrained value type.
 template <typename T>
-using NonNullFF = std::enable_if_t<std::is_pointer<T>::value, Checked<T, NonZeroChecker<T>, true>>;
+using NonNullFF = std::enable_if_t<std::is_pointer<T>::value, detail::Checked<T, detail::NonZeroChecker<T>, true>>;
 
 static_assert(!std::is_default_constructible<NonNull<int*>>::value);
