+++ Library/include/playrho/detail/NonPositiveChecker.hpp
@@ -18,25 +18,34 @@
  * 3. This notice may not be removed or altered from any source distribution.
  */
 
-#ifndef PLAYRHO_UNITINTERVAL_HPP
-#define PLAYRHO_UNITINTERVAL_HPP
+#ifndef PLAYRHO_DETAIL_NONPOSITIVECHECKER_HPP
+#define PLAYRHO_DETAIL_NONPOSITIVECHECKER_HPP
 
-#include <playrho/detail/UnitIntervalChecker.hpp>
+#include <playrho/detail/Checked.hpp>
 
-namespace playrho {
+namespace playrho::detail {
 
-/// @ingroup CheckedTypes
-/// @brief Unit interval constrained value type.
+/// @brief Non-positive constrained value checker.
 template <typename T>
-using UnitInterval = detail::Checked<T, detail::UnitIntervalChecker<T>>;
+struct NonPositiveChecker {
 
-/// @ingroup CheckedTypes
-/// @brief Fast failing unit interval constrained value type.
-template <typename T>
-using UnitIntervalFF = detail::Checked<T, detail::UnitIntervalChecker<T>, true>;
+    /// @brief Default value supplying functor.
+    constexpr auto operator()() noexcept -> decltype(static_cast<T>(0))
+    {
+        return static_cast<T>(0);
+    }
 
-static_assert(std::is_default_constructible<UnitInterval<int>>::value);
+    /// @brief Value checking functor.
+    constexpr auto operator()(const T& v) noexcept
+        -> decltype(v <= static_cast<T>(0), static_cast<const char*>(nullptr))
+    {
+        if (!(v <= static_cast<T>(0))) {
+            return "value not lesser than nor equal to zero";
+        }
+        return {};
+    }
+};
 
-} // namespace playrho
+} // namespace playrho::detail
 
-#endif // PLAYRHO_UNITINTERVAL_HPP
+#endif // PLAYRHO_DETAIL_NONPOSITIVECHECKER_HPP
